fix jsutil for support GopherJS/wasm Go 1.2/1.3/1.4
