+++ internal/jsutil/go112_js.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// +build !go1.13 !wasm
+// +build !go1.13
 
 package jsutil
